+++ client/actions/actionTypes.js
@@ -10,7 +10,5 @@ export const ADD_COMPETITOR_DATA = 'ADD_COMPETITOR_DATA';
 export const ADD_DAILY_COMPETITOR_DATA = 'ADD_DAILY_COMPETITOR_DATA';
 export const ADD_COMPETITOR = 'ADD_COMPETITOR';
 export const CHOOSE_WEAPON = 'CHOOSE_WEAPON';
-export const REFRESH_FRIENDS = 'REFRESH_FRIENDS';
-export const REFRESH_FRIENDS = 'REFRESH_FRIENDS';
 export const RECEIVED_FR = 'RECEIVED_FR';
 export const SENT_FR = 'SENT_FR';
